scripts/requirements.txt
requests==2.32.5
beautifulsoup4==4.14.2
markdownify==1.2.2
python-dotenv==1.2.1
